requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.0.3
textdistance==4.6.2
PyPDF2==3.0.1
python-docx==0.8.11
textstat==0.7.3
python-magic==0.4.27